requirements.txt
appdirs==1.4.4
astroid==2.11.2
atomicwrites==1.4.0
attrs==21.4.0
audioread==2.1.9
backcall==0.2.0
bidict==0.22.0
black==22.3.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.1.2
colorama==0.4.4
crypto==1.4.1
cryptography==36.0.2
cycler==0.10.0
decorator==5.0.9
dill==0.3.4
flake8==4.0.1
Flask==2.1.1
Flask-SocketIO==5.1.1
idna==2.10
importlib-metadata==4.11.3
imutils==0.5.4
iniconfig==1.1.1
ipython==7.25.0
ipython-genutils==0.2.0
isort==5.10.1
itsdangerous==2.1.2
jedi==0.18.0
Jinja2==3.1.1
joblib==1.0.1
keyboard==0.13.5
kiwisolver==1.3.1
lazy-object-proxy==1.7.1
librosa==0.8.1
llvmlite==0.36.0
lxml==4.8.0
MarkupSafe==2.1.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mccabe==0.6.1
MouseInfo==0.1.3
mypy-extensions==0.4.3
Naked==0.1.31
numba==0.53.1
numpy==1.21.0
opcua==0.98.13
opencv-python==4.5.2.54
packaging==21.0
pandas==1.2.5
parso==0.8.2
pathspec==0.9.0
pickleshare==0.7.5
Pillow==8.2.0
platformdirs==2.5.2
pluggy==1.0.0
pooch==1.4.0
prompt-toolkit==3.0.19
py==1.11.0
PyAutoGUI==0.9.52
pycodestyle==2.8.0
pycparser==2.20
pyflakes==2.4.0
PyGetWindow==0.0.9
Pygments==2.9.0
pylint==2.13.5
PyMsgBox==1.0.9
pyodbc==4.0.30
pyparsing==2.4.7
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.27
pytest==7.1.1
python-dateutil==2.8.1
python-engineio==4.3.1
python-socketio==5.5.2
PyTweening==1.0.3
pytz==2021.1
PyYAML==6.0
requests==2.25.1
resampy==0.2.2
scikit-learn==0.24.2
scipy==1.7.0
shellescape==3.8.1
six==1.16.0
SoundFile==0.10.3.post1
threadpoolctl==2.1.0
tomli==2.0.1
traitlets==5.0.5
typing_extensions==4.2.0
urllib3==1.26.6
wcwidth==0.2.5
Werkzeug==2.1.1
wrapt==1.14.0
zipp==3.8.0
